Creando una rama de pruebas del proyecto completo

+++ src/main/java/com/intellimarket/api/inventory/dto/ProductsResponse.java
@@ -9,5 +9,5 @@ public record ProductsResponse (
         String description,
         Integer stock,
         BigDecimal price,
-        BigInteger id
+        Long id
 ) {}
